% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{BayFDR}
\alias{BayFDR}
\title{function of BFDR}
\usage{
BayFDR(PPI, alpha)
}
\arguments{
\item{PPI}{:the sequence of the result}

\item{alpha}{:the significance level}
}
\value{
：the threshold determining the SV and non-SV genes
}
\description{
function of BFDR
}
