% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{compu_phi}
\alias{compu_phi}
\title{function of Numerical integration2}
\usage{
compu_phi(n = 200, s, t, ap, max_attempts = 10)
}
\arguments{
\item{n}{:parameters of algorithm}

\item{s}{:parameters of function}

\item{t}{:parameters of function}

\item{ap}{:parameters of function}

\item{max_attempts}{:max attempts}
}
\value{
result of the Numerical integration
}
\description{
function of Numerical integration2
}
