% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{compute_diag_ABC_corrected}
\alias{compute_diag_ABC_corrected}
\title{function of matrix compution}
\usage{
compute_diag_ABC_corrected(A, B)
}
\arguments{
\item{A}{:Matrix1}

\item{B}{:Matrix2}
}
\value{
result of multi-matrix compution
}
\description{
function of matrix compution
}
