% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction0806.R
\name{CTIG}
\alias{CTIG}
\title{data preprocess}
\usage{
CTIG(spelist)
}
\arguments{
\item{spelist}{A list of length m. Each element is a sublist containing:
\itemize{
\item{expression_matrix: Matrix (spots x genes)}
\item{location_matrix: Matrix (spots x coordinates)}
}}
}
\value{
: preprocessed data
}
\description{
data preprocess
}
