% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction0806.R
\name{f}
\alias{f}
\title{function to compute complex function}
\usage{
f(x, p, q, r, s, t, h = 0)
}
\arguments{
\item{x}{:value}

\item{p}{:parameters of function}

\item{q}{:parameters of function}

\item{r}{:parameters of function}

\item{s}{:parameters of function}

\item{t}{:parameters of function}

\item{h}{:parameters of function}
}
\value{
value of the function f(x)
}
\description{
function to compute complex function
}
