% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction0806.R
\name{gausslq}
\alias{gausslq}
\title{function of Numerical integration}
\usage{
gausslq(n = 100, p, q, r, s, t, h = 0)
}
\arguments{
\item{n}{:parameters of algorithm}

\item{p}{:parameters of function}

\item{q}{:parameters of function}

\item{r}{:parameters of function}

\item{s}{:parameters of function}

\item{t}{:parameters of function}

\item{h}{:parameters of function}
}
\value{
result of the Numerical integration
}
\description{
function of Numerical integration
}
