% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{log_compu_phi}
\alias{log_compu_phi}
\title{function of Numerical integration4}
\usage{
log_compu_phi(s, t, ap)
}
\arguments{
\item{s}{:parameters of function}

\item{t}{:parameters of function}

\item{ap}{:parameters of function}
}
\value{
result of the Numerical integration
}
\description{
function of Numerical integration4
}
