% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{safe_compu_phi}
\alias{safe_compu_phi}
\title{function of Numerical integration3}
\usage{
safe_compu_phi(n = 200, s, t, ap, timeout = 0.05, old = 1)
}
\arguments{
\item{n}{:parameters of algorithm}

\item{s}{:parameters of function}

\item{t}{:parameters of function}

\item{ap}{:parameters of function}

\item{timeout}{:the maximum process time}

\item{old}{: post value of parameter}
}
\value{
result of the Numerical integration
}
\description{
function of Numerical integration3
}
