% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction0806.R
\name{safe_log_compu_phi}
\alias{safe_log_compu_phi}
\title{function of Numerical integration5}
\usage{
safe_log_compu_phi(s, t, ap, timeout = 0.05, old = 0)
}
\arguments{
\item{s}{:parameters of function}

\item{t}{:parameters of function}

\item{ap}{:parameters of function}

\item{timeout}{:the maximum process time}

\item{old}{: post value of parameter}
}
\value{
result of the Numerical integration
}
\description{
function of Numerical integration5
}
