% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction0806.R
\name{sim_create}
\alias{sim_create}
\title{function to produce simulation}
\usage{
sim_create(
  gene_size = 100,
  svgene_size = 0.1,
  sv_mark = c(1, 1),
  no_sv_mark = c(0, 0),
  inf_size = 0.1,
  phi = 15,
  etamean = 2,
  xspace = "linear",
  yspace = "linear",
  seed = 1,
  cell_dist = rep(1, 6),
  domainnum = 5
)
}
\arguments{
\item{gene_size}{:the number of gene numbers you need}

\item{svgene_size}{:the proportion of SV gene among the gene_size you need}

\item{sv_mark}{:the coefficients of spatial effect for SV genes in 2-dimention direction you need}

\item{no_sv_mark}{:the coefficients of spatial effect for non-SV genes in 2-dimention direction you need.Default is 0.}

\item{inf_size}{:the probability of dropout}

\item{phi}{:the shape parameter of NB distribution}

\item{etamean}{:the mean value of the baseline eta.Default is 2.}

\item{xspace}{:the spatial pattern in x coordinate.Default situations are "linear","focal" and "period".}

\item{yspace}{:the spatial pattern in x coordinate.Default situations are "linear","focal" and "period".}

\item{seed}{:seed number}

\item{cell_dist}{:the alpha of the Dirichlet distribution of the coefficients of covariates. We default the covariate's number =6.}

\item{domainnum}{:different covariate domain structures. The value domain is from 1 to 11 and we default it to 5.}
}
\value{
a list consist of expression matrix, spot location and covariate matrix.
}
\description{
function to produce simulation
}
