% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{tun_spl}
\alias{tun_spl}
\title{choose the basic function degree}
\usage{
tun_spl(spe.list, calpha.list, g = 1)
}
\arguments{
\item{spe.list}{:same to spelist}

\item{calpha.list}{:same to c_alpha}

\item{g}{:gene number}
}
\value{
the number of basic function degree
}
\description{
choose the basic function degree
}
