% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{VIZINB}
\alias{VIZINB}
\title{implement function}
\usage{
VIZINB(
  y,
  splinevalue,
  samplenumber,
  splinelevel,
  acrate = 0.01,
  ak = 1,
  dp = 1.8,
  cp = 0.2,
  c_alpha,
  knum = 6,
  cq = 1,
  dq = 1,
  gamma1 = 0.01,
  gamma2 = 0.01
)
}
\arguments{
\item{y}{：list consist of a gene expression in m datasets}

\item{splinevalue}{:list consist of basic functions in m datasets with predetermined degree}

\item{samplenumber}{:vector consist of gene numbers in m datasets}

\item{splinelevel}{:predetermined degree}

\item{acrate}{:threshold value for determining ELBO convergence}

\item{ak}{:hyperparameter of varience of beta to control the slab and spike prior}

\item{dp}{:hyperparameter of alpha's beta prior}

\item{cp}{:hyperparameter of alpha's beta prior}

\item{c_alpha}{:list consist of coveriates in m datasets}

\item{knum}{:the dimension of coveriates}

\item{cq}{:hyperparameter of u's beta prior}

\item{dq}{:hyperparameter of u's beta prior}

\item{gamma1}{:hyperparameter of the slab and spike prior of beta}

\item{gamma2}{:hyperparameter of the bi-level structure of alpha}
}
\value{
the posterior mean of u
}
\description{
implement function
}
