% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction0806.R
\name{citgtest}
\alias{citgtest}
\title{single gene process}
\usage{
citgtest(
  g,
  result,
  c_alpha,
  spelist,
  acrate,
  gamma1,
  gamma2,
  min_iter,
  max_iter,
  speci_iter
)
}
\arguments{
\item{g}{:gene number}

\item{result}{: preprocessed data produced by CTIG}

\item{c_alpha}{A list of length m, each element being a covariates matrix.}

\item{spelist}{A list of length m. Each element is a sublist containing:
\itemize{
\item{expression_matrix: Matrix (spots x genes)}
\item{location_matrix: Matrix (spots x coordinates)}
}}

\item{acrate}{:the minimum threshold for EBLO convergence.}

\item{gamma1}{:hyperparameter of the slab and spike prior of beta. Default is sqrt(0.001).}

\item{gamma2}{:hyperparameter of the bi-level structure of alpha. We set gamma2 to 0.01, 0.005, and 0.001 for more than four, three, and two or fewer slices, respectively.}

\item{min_iter}{: the minimum number of iteration.}

\item{max_iter}{: the maximum number of iteration.}

\item{speci_iter}{: Number of times the early convergence criterion is met.}
}
\value{
the posterior mean of u_k of gene g
}
\description{
single gene process
}
