% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{tun_spl}
\alias{tun_spl}
\title{choose the basic function degree}
\usage{
tun_spl(spelist, calpha.list, g = 1)
}
\arguments{
\item{spelist}{A list of length m. Each element is a sublist containing:
\itemize{
\item{expression_matrix: Matrix (spots x genes)}
\item{location_matrix: Matrix (spots x  coordinates)}
}}

\item{calpha.list}{:same to c_alpha}

\item{g}{:gene number}
}
\value{
the number of basic function degree
}
\description{
choose the basic function degree
}
