% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{validate_data_consistency_strict}
\alias{validate_data_consistency_strict}
\title{data Detection}
\usage{
validate_data_consistency_strict(spe.list, c_alpha)
}
\arguments{
\item{c_alpha}{A list of length m, each element being a covariates matrix.}

\item{spelist}{A list of length m. Each element is a sublist containing:
\itemize{
\item{expression_matrix: Matrix (spots x genes)}
\item{location_matrix: Matrix (spots x \link{x, y} coordinates)}
}}
}
\value{
TRUE or warnings
}
\description{
data Detection
}
