% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction0806.R
\name{VIZINB}
\alias{VIZINB}
\title{implement function}
\usage{
VIZINB(
  y,
  splinevalue,
  samplenumber,
  splinelevel,
  acrate = 0.01,
  ak = 1,
  dp = 1.8,
  cp = 0.2,
  c_alpha,
  knum = 6,
  cq = 1,
  dq = 1,
  gamma1 = sqrt(0.001),
  gamma2 = 0.01,
  min_iter = 30,
  max_iter = 200,
  speci_iter = 12
)
}
\arguments{
\item{y}{：list consist of a gene expression in m datasets}

\item{splinevalue}{:list consist of basic functions in m datasets with predetermined degree}

\item{samplenumber}{:vector consist of gene numbers in m datasets}

\item{splinelevel}{:predetermined degree}

\item{acrate}{:threshold value for determining ELBO convergence}

\item{ak}{:hyperparameter of varience of beta to control the slab and spike prior}

\item{dp}{:hyperparameter of alpha's beta prior}

\item{cp}{:hyperparameter of alpha's beta prior}

\item{c_alpha}{:list consist of coveriates in m datasets}

\item{knum}{:the dimension of coveriates}

\item{cq}{:hyperparameter of u's beta prior}

\item{dq}{:hyperparameter of u's beta prior}

\item{gamma1}{:hyperparameter of the slab and spike prior of beta. Default is sqrt(0.001).}

\item{gamma2}{:hyperparameter of the bi-level structure of alpha. We set gamma2 to 0.01, 0.005, and 0.001 to represent scenarios integrating more than four, three, and two or fewer slices, respectively.}

\item{min_iter}{: the minimum number of iteration.}

\item{max_iter}{: the maximum number of iteration.}

\item{speci_iter}{: Number of times the early convergence criterion is met}
}
\value{
the posterior mean of u
}
\description{
implement function
}
